chore #16 누락 필드 및 타입 변경

+++ src/main/java/land/leets/Carrot/domain/career/repository/WorkTypeRepository.java
@@ -4,6 +4,6 @@ import java.util.Optional;
 import land.leets.Carrot.domain.career.entity.WorkType;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface WorkTypeRepository extends JpaRepository<WorkType, Long> {
+public interface WorkTypeRepository extends JpaRepository<WorkType, Integer> {
     public Optional<WorkType> findByType(String typeName);
 }
